requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
python-dotenv==1.0.1
openai==1.12.0
langdetect==1.0.9
beautifulsoup4==4.12.3
python-mpesa==0.1.10
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
locust==2.24.0
sentry-sdk==1.39.1
gunicorn==21.2.0
python-multipart==0.0.9
requests>=2.26.0 